chore: remove dependency on flink sql language service (#85)
We can't put this into dependencies because it's a close source
thing that we depend on in OSS thing (cli)
